+++ class/scrumcard.class.php
@@ -1081,6 +1081,7 @@ class ScrumCard extends CommonObject
 		$object->type = 'scrum-card';
 		$object->class = array();     // array of additional classes
 		$object->element = $this->element;
+		$object->socid = 0;
 		$object->cardUrl = dol_buildpath('/scrumproject/scrumcard_card.php',1).'?id='.$this->id;
 		$object->objectId = $this->id;
 		$object->title = '';
@@ -1140,16 +1141,7 @@ class ScrumCard extends CommonObject
 				$us = scrumProjectGetObjectByElement('scrumproject_scrumuserstory', $elementObject->fk_scrum_user_story);
 
 				if($us ){
-					$object->label = $us->label;
-
-//					/** @var Task $elementObject */
-//					$task = scrumProjectGetObjectByElement('task', $us->fk_task);
-//					if($task){
-//						$object->label = $task->label; // les us plannifiées n'ont pas de libellé
-//					}
-//					else{
-//						$object->label = '<span class="error">Task Error</span>';
-//					}
+					$object->label = !empty($elementObject->label)?$elementObject->label:$us->label;
 				}
 				else{
 					$object->label = '<span class="error">US Error</span>';
@@ -1200,6 +1192,14 @@ class ScrumCard extends CommonObject
 
 
 		$object->title.= '<div class="kanban-item__body">';
+
+		if(!empty($object->socid)){
+			$company = new Societe($this->db);
+			if($company->fetch($object->socid)>0){
+				$object->title.= '<span class="kanban-item__company">'.$company->name.'</span>';
+			}
+		}
+
 		$object->title.= '<span class="kanban-item__label">'.$object->label.'</span>';
 		$object->title.= '</div>';