+++ ChangeLog.md
@@ -2,11 +2,20 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+
+
+## Version 1.24
+
+- FIX : Can't split card from kanban when overspend *07/01/2023* - 1.24.3
+- FIX : Missing substitution *04/01/2022* - 1.24.2
+- FIX : Live update for time qty in text *02/01/2022* - 1.24.1
+- NEW : Display company on card *02/01/2022* - 1.24.0
 - NEW : Delete cascade from kanban *20/12/2022* - 1.23.0
 
 ## Version 1.22
 
-- FIX : Can't split card from kanban when overspend *07/01/2023* - 1.22.1
+- FIX : Us sprint label display *02/01/2023* - 1.22.3
 - FIX : Can't validate a scrumsprintuser card beacause ref column doesn't exist for table scrumproject_scrumsprintuser *23/12/2022* - 1.22.2
 - FIX : fix scrumuserstorysprint_import_wizard page filter columns *07/12/2022* - 1.22.1
 - NEW : Add substitution *30/11/2022* - 1.22.0